+++ sdk/python/lib/pulumi/runtime/settings.py
@@ -328,21 +328,15 @@ async def monitor_supports_alias_specs() -> bool:
 
 
 def _sync_monitor_supports_transforms() -> bool:
-    if "transforms" not in SETTINGS.feature_support:
-        return False
-    return SETTINGS.feature_support["transforms"]
+    return SETTINGS.feature_support.get("transforms", False)
 
 
 def _sync_monitor_supports_invoke_transforms() -> bool:
-    if "invokeTransforms" not in SETTINGS.feature_support:
-        return False
-    return SETTINGS.feature_support["invokeTransforms"]
+    return SETTINGS.feature_support.get("invokeTransforms", False)
 
 
 def _sync_monitor_supports_parameterization() -> bool:
-    if "parameterization" not in SETTINGS.feature_support:
-        return False
-    return SETTINGS.feature_support["parameterization"]
+    return SETTINGS.feature_support.get("parameterization", False)
 
 
 def reset_options(