+++ tests/testdata/codegen/plain-object-defaults/python/pulumi_example/mod1/_inputs.py
@@ -4,14 +4,31 @@
 
 import copy
 import warnings
+import sys
 import pulumi
 import pulumi.runtime
 from typing import Any, Mapping, Optional, Sequence, Union, overload
+if sys.version_info >= (3, 11):
+    from typing import NotRequired, TypedDict, TypeAlias
+else:
+    from typing_extensions import NotRequired, TypedDict, TypeAlias
 from .. import _utilities
 
 __all__ = [
     'TypArgs',
+    'TypArgsDict',
 ]
+
+MYPY = False
+
+if not MYPY:
+    class TypArgsDict(TypedDict):
+        """
+        A test for namespaces (mod 1)
+        """
+        val: NotRequired[pulumi.Input[str]]
+elif False:
+    TypArgsDict: TypeAlias = Mapping[str, Any]
 
 @pulumi.input_type
 class TypArgs: